Move timestamp lexer into tokenize directory
